migration from `click.argument` to `click.prompt`
